+++ apps/extension/client/src/constants.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 
 export const getBaseUrl = (): string =>
-  vscode.workspace.getConfiguration().get("xstate.targetEditorBaseUrl");
+  vscode.workspace.getConfiguration().get("xstate.targetEditorBaseUrl")!;
 export const getTokenKey = () => {
   const baseUrl = getBaseUrl();
   return `stately-editor-key#${baseUrl}`;