+++ apps/extension/client/src/handleDefinitionUpdate.ts
@@ -1,4 +1,3 @@
-import * as vscode from "vscode";
 import { parseMachinesFromFile } from "@xstate/machine-extractor";
 import {
   getInlineImplementations,
@@ -6,12 +5,13 @@ import {
   getRangeFromSourceLocation,
   resolveUriToFilePrefix,
 } from "@xstate/tools-shared";
+import * as vscode from "vscode";
 import { UpdateDefinitionEvent } from "./editorWebviewScript";
 
 export const handleDefinitionUpdate = async (event: UpdateDefinitionEvent) => {
   const doc = vscode.workspace.textDocuments.find((doc) => {
     return resolveUriToFilePrefix(doc.uri.path) === event.uri;
-  });
+  })!;
 
   const text = doc.getText();
 
@@ -23,13 +23,13 @@ export const handleDefinitionUpdate = async (event: UpdateDefinitionEvent) => {
 
   const workspaceEdit = new vscode.WorkspaceEdit();
 
-  const range = getRangeFromSourceLocation(machine.ast.definition?.node?.loc);
+  const range = getRangeFromSourceLocation(machine.ast.definition?.node?.loc!);
 
   workspaceEdit.replace(
     doc.uri,
     new vscode.Range(
       new vscode.Position(range.start.line, range.start.character),
-      new vscode.Position(range.end.line, range.end.character),
+      new vscode.Position(range.end.line, range.end.character)
     ),
     await getNewMachineText({
       fileName: doc.fileName,
@@ -37,7 +37,7 @@ export const handleDefinitionUpdate = async (event: UpdateDefinitionEvent) => {
       implementations: getInlineImplementations(machine, text),
       text,
       newConfig: event.config,
-    }),
+    })
   );
 
   const layoutComment = machine.getLayoutComment()?.comment;
@@ -50,18 +50,18 @@ export const handleDefinitionUpdate = async (event: UpdateDefinitionEvent) => {
       doc.uri,
       new vscode.Range(
         new vscode.Position(range.start.line, range.start.character),
-        new vscode.Position(range.end.line, range.end.character),
+        new vscode.Position(range.end.line, range.end.character)
       ),
-      `/** @xstate-layout ${event.layoutString} */`,
+      `/** @xstate-layout ${event.layoutString} */`
     );
   } else {
     // Insert layout comment
-    const range = getRangeFromSourceLocation(machine.ast.callee.loc);
+    const range = getRangeFromSourceLocation(machine.ast.callee.loc!);
 
     workspaceEdit.insert(
       doc.uri,
       new vscode.Position(range.start.line, range.start.character),
-      `\n/** @xstate-layout ${event.layoutString} */\n`,
+      `\n/** @xstate-layout ${event.layoutString} */\n`
     );
   }