+++ packages/machine-extractor/src/schema.ts
@@ -1,5 +1,4 @@
 import { AnyNode } from "./scalars";
-import { unionType } from "./unionType";
 import { objectTypeWithKnownKeys } from "./utils";
 
 export const Schema = objectTypeWithKnownKeys({