+++ packages/shared/src/__tests__/getTypegenOutput.test.ts
@@ -1,5 +1,6 @@
 import { execSync } from "child_process";
 import * as fs from "fs";
+import * as minimatch from "minimatch";
 import * as path from "path";
 import {
   getDocumentValidationsResults,
@@ -8,33 +9,35 @@ import {
 } from "..";
 
 describe("getTypegenOutput", () => {
-  execSync("rm -rf ./__examples__/*.typegen.ts", {
-    cwd: __dirname,
-  });
+  const examplesPath = path.resolve(__dirname, "__examples__");
 
-  const dir = fs.readdirSync(path.resolve(__dirname, "__examples__"));
+  minimatch
+    .match(fs.readdirSync(examplesPath), "*.typegen.ts")
+    .map((file) => fs.unlinkSync(path.join(examplesPath, file)));
 
-  const tsExtensionFiles = dir.filter((file) => file.endsWith(".ts"));
+  const tsExtensionFiles = fs
+    .readdirSync(examplesPath)
+    .filter((file) => file.endsWith(".ts"));
 
   tsExtensionFiles.forEach((file) => {
     const fileText = fs.readFileSync(
       path.resolve(__dirname, "__examples__", file),
-      "utf8",
+      "utf8"
     );
 
     const event = makeXStateUpdateEvent(
       // URI doesn't matter here
       "",
-      getDocumentValidationsResults(fileText),
+      getDocumentValidationsResults(fileText)
     );
 
     fs.writeFileSync(
       path.resolve(
         __dirname,
         "__examples__",
-        file.slice(0, -3) + ".typegen.ts",
+        file.slice(0, -3) + ".typegen.ts"
       ),
-      getTypegenOutput(event),
+      getTypegenOutput(event)
     );
   });